+++ src/onegov.gazette/onegov/gazette/upgrade.py
@@ -3,61 +3,7 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 from onegov.core.upgrade import upgrade_task
-from onegov.gazette.collections.categories import CategoryCollection
-from onegov.gazette.collections.organizations import OrganizationCollection
 from onegov.gazette.models.notice import GazetteNotice
-
-
-@upgrade_task('Migrate gazette categories', always_run=True)
-def migrate_categories(context):
-    principal = getattr(context.app, 'principal', None)
-    if not principal:
-        return False
-
-    categories = getattr(principal, '_categories', None)
-    if not categories:
-        return False
-
-    if not context.has_table('gazette_categories'):
-        return False
-
-    session = context.app.session_manager.session()
-    count = session.execute("select count(*) from gazette_categories")
-    if count.scalar() != 0:
-        return False
-
-    collection = CategoryCollection(session)
-    for name, title in categories.items():
-        collection.add_root(name=name, title=title, active=True)
-
-
-@upgrade_task('Migrate gazette organizations', always_run=True)
-def migrate_organizations(context):
-    principal = getattr(context.app, 'principal', None)
-    if not principal:
-        return False
-
-    organizations = getattr(principal, '_organizations', None)
-    if not organizations:
-        return False
-
-    if not context.has_table('gazette_organizations'):
-        return False
-
-    session = context.app.session_manager.session()
-    session.execute("delete from gazette_organizations")  # todo: remove me!!!!
-    count = session.execute("select count(*) from gazette_organizations")
-    if count.scalar() != 0:
-        return False
-
-    collection = OrganizationCollection(session)
-    for index, name in enumerate(organizations):
-        collection.add_root(
-            name=name,
-            title=organizations[name],
-            active=True,
-            order=index
-        )
 
 
 @upgrade_task(